requirements.txt
# Supabase client
supabase>=2.0.0

# HTTP client
httpx[http2]>=0.28.0

# Browser automation
playwright>=1.49.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Environment
python-dotenv>=1.0.0

# Data
pandas>=2.0.0

# Email (Phase 5)
resend>=2.0.0
